Remove Java NIO frame handler
Replaced by Netty-based frame handler.
