--- src/main/java/com/rabbitmq/client/impl/nio/FrameWriteRequest.java
@@ -1,38 +1,0 @@
-// Copyright (c) 2007-2025 Broadcom. All Rights Reserved. The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
-//
-// This software, the RabbitMQ Java client library, is triple-licensed under the
-// Mozilla Public License 2.0 ("MPL"), the GNU General Public License version 2
-// ("GPL") and the Apache License version 2 ("ASL"). For the MPL, please see
-// LICENSE-MPL-RabbitMQ. For the GPL, please see LICENSE-GPL2.  For the ASL,
-// please see LICENSE-APACHE2.
-//
-// This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND,
-// either express or implied. See the LICENSE file for specific language governing
-// rights and limitations of this software.
-//
-// If you have any questions regarding licensing, please contact us at
-// info@rabbitmq.com.
-
-package com.rabbitmq.client.impl.nio;
-
-import com.rabbitmq.client.impl.Frame;
-
-import java.io.DataOutputStream;
-import java.io.IOException;
-
-/**
- *
- */
-public class FrameWriteRequest implements WriteRequest {
-
-    final Frame frame;
-
-    public FrameWriteRequest(Frame frame) {
-        this.frame = frame;
-    }
-
-    @Override
-    public void handle(DataOutputStream outputStream) throws IOException {
-        frame.writeTo(outputStream);
-    }
-}
