--- src/main/java/com/rabbitmq/client/impl/nio/package-info.java
@@ -1,4 +1,0 @@
-/**
- * NIO network connector.
- */
-package com.rabbitmq.client.impl.nio;
