+++ src/test/java/com/rabbitmq/client/test/ClientTestSuite.java
@@ -46,14 +46,12 @@ import org.junit.platform.suite.api.Suite;
     MetricsCollectorTest.class,
     MicrometerMetricsCollectorTest.class,
     DnsSrvRecordAddressResolverTest.class,
-    JavaNioTest.class,
     ConnectionFactoryTest.class,
     RecoveryAwareAMQConnectionFactoryTest.class,
     RpcTest.class,
     LambdaCallbackTest.class,
     ChannelAsyncCompletableFutureTest.class,
     RecoveryDelayHandlerTest.class,
-    FrameBuilderTest.class,
     PropertyFileInitialisationTest.class,
     ClientVersionTest.class,
     TestUtilsTest.class,
@@ -62,7 +60,6 @@ import org.junit.platform.suite.api.Suite;
     JacksonJsonRpcTest.class,
     AddressTest.class,
     DefaultRetryHandlerTest.class,
-    NioDeadlockOnConnectionClosing.class,
     GeneratedClassesTest.class,
     RpcTopologyRecordingTest.class,
     ConnectionTest.class,
