--- src/test/java/com/rabbitmq/client/test/JavaNioTest.java
@@ -1,251 +1,0 @@
-package com.rabbitmq.client.test;
-
-import com.rabbitmq.client.*;
-import com.rabbitmq.client.impl.nio.BlockingQueueNioQueue;
-import com.rabbitmq.client.impl.nio.DefaultByteBufferFactory;
-import com.rabbitmq.client.impl.nio.NioParams;
-import org.assertj.core.api.Condition;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.util.List;
-import java.util.concurrent.*;
-import java.util.concurrent.atomic.AtomicInteger;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-/**
- *
- */
-public class JavaNioTest {
-
-    public static final String QUEUE = "nio.queue";
-
-    private Connection testConnection;
-
-    @BeforeEach
-    public void init() throws Exception {
-        ConnectionFactory connectionFactory = new ConnectionFactory()
-                .useNio();
-        testConnection = connectionFactory.newConnection();
-    }
-
-    @AfterEach
-    public void tearDown() throws Exception {
-        if (testConnection != null) {
-            testConnection.createChannel().queueDelete(QUEUE);
-            testConnection.close();
-        }
-    }
-
-    @Test
-    public void connection() throws Exception {
-        CountDownLatch latch = new CountDownLatch(1);
-        ConnectionFactory connectionFactory = new ConnectionFactory()
-                .useNio();
-        Connection connection = null;
-        try {
-            connection = basicGetBasicConsume(connectionFactory, "nio.queue", latch);
-            boolean messagesReceived = latch.await(5, TimeUnit.SECONDS);
-            assertTrue(messagesReceived, "Message has not been received");
-        } finally {
-            safeClose(connection);
-        }
-    }
-
-    @Test
-    public void twoConnections() throws IOException, TimeoutException, InterruptedException {
-        CountDownLatch latch = new CountDownLatch(2);
-        ConnectionFactory connectionFactory = new ConnectionFactory()
-                .useNio()
-                .setNioParams(new NioParams().setNbIoThreads(4));
-        Connection connection1 = null;
-        Connection connection2 = null;
-        try {
-            connection1 = basicGetBasicConsume(connectionFactory, "nio.queue.1", latch);
-            connection2 = basicGetBasicConsume(connectionFactory, "nio.queue.2", latch);
-
-            boolean messagesReceived = latch.await(5, TimeUnit.SECONDS);
-            assertTrue(messagesReceived, "Messages have not been received");
-        } finally {
-            safeClose(connection1);
-            safeClose(connection2);
-        }
-    }
-
-    @Test
-    public void twoConnectionsWithNioExecutor() throws IOException, TimeoutException, InterruptedException {
-        CountDownLatch latch = new CountDownLatch(2);
-        ExecutorService nioExecutor = Executors.newFixedThreadPool(5);
-        ConnectionFactory connectionFactory = new ConnectionFactory()
-                .useNio();
-        Connection connection1 = null;
-        Connection connection2 = null;
-        try {
-            connection1 = basicGetBasicConsume(connectionFactory, "nio.queue.1", latch);
-            connection2 = basicGetBasicConsume(connectionFactory, "nio.queue.2", latch);
-
-            boolean messagesReceived = latch.await(5, TimeUnit.SECONDS);
-            assertTrue(messagesReceived, "Messages have not been received");
-        } finally {
-            safeClose(connection1);
-            safeClose(connection2);
-            nioExecutor.shutdownNow();
-        }
-    }
-
-    @Test
-    public void shutdownListenerCalled() throws IOException, TimeoutException, InterruptedException {
-        ConnectionFactory connectionFactory = new ConnectionFactory()
-                .useNio();
-        Connection connection = connectionFactory.newConnection();
-        try {
-            final CountDownLatch latch = new CountDownLatch(1);
-            connection.addShutdownListener(new ShutdownListener() {
-
-                @Override
-                public void shutdownCompleted(ShutdownSignalException cause) {
-                    latch.countDown();
-                }
-            });
-            safeClose(connection);
-            assertTrue(latch.await(5, TimeUnit.SECONDS), "Shutdown listener should have been called");
-        } finally {
-            safeClose(connection);
-        }
-    }
-
-    @Test
-    public void nioLoopCleaning() throws Exception {
-        ConnectionFactory connectionFactory = new ConnectionFactory()
-                .useNio();
-        for (int i = 0; i < 10; i++) {
-            Connection connection = connectionFactory.newConnection();
-            connection.abort();
-        }
-    }
-
-    @Test
-    public void messageSize() throws Exception {
-        for (int i = 0; i < 50; i++) {
-            sendAndVerifyMessage(testConnection, 76390);
-        }
-    }
-
-    @Test
-    public void byteBufferFactory() throws Exception {
-        ConnectionFactory connectionFactory = new ConnectionFactory()
-                .useNio();
-        int baseCapacity = 32768;
-        NioParams nioParams = new NioParams();
-        nioParams.setReadByteBufferSize(baseCapacity / 2);
-        nioParams.setWriteByteBufferSize(baseCapacity / 4);
-        List<ByteBuffer> byteBuffers = new CopyOnWriteArrayList<>();
-        connectionFactory.setNioParams(nioParams.setByteBufferFactory(new DefaultByteBufferFactory(capacity -> {
-            ByteBuffer bb = ByteBuffer.allocate(capacity);
-            byteBuffers.add(bb);
-            return bb;
-        })));
-
-        try (Connection c = connectionFactory.newConnection()) {
-            sendAndVerifyMessage(c, 100);
-        }
-
-        assertThat(byteBuffers).hasSize(2);
-        Condition<Integer> condition = new Condition<>(c -> c == nioParams.getReadByteBufferSize() ||
-                c == nioParams.getWriteByteBufferSize(), "capacity set by factory");
-        assertThat(byteBuffers.get(0).capacity()).is(condition);
-        assertThat(byteBuffers.get(1).capacity()).is(condition);
-    }
-
-    @Test
-    public void directByteBuffers() throws Exception {
-        ConnectionFactory connectionFactory = new ConnectionFactory()
-                .useNio()
-                .setNioParams(new NioParams().setByteBufferFactory(new DefaultByteBufferFactory(capacity -> ByteBuffer.allocateDirect(capacity))));
-        try (Connection c = connectionFactory.newConnection()) {
-            sendAndVerifyMessage(c, 100);
-        }
-    }
-
-    @Test
-    public void customWriteQueue() throws Exception {
-        AtomicInteger count = new AtomicInteger(0);
-        ConnectionFactory connectionFactory = new ConnectionFactory()
-                .useNio()
-                .setNioParams(new NioParams().setWriteQueueFactory(ctx -> {
-                    count.incrementAndGet();
-                    return new BlockingQueueNioQueue(
-                            new LinkedBlockingQueue<>(ctx.getNioParams().getWriteQueueCapacity()),
-                            ctx.getNioParams().getWriteEnqueuingTimeoutInMs()
-                    );
-                }));
-        try (Connection c = connectionFactory.newConnection()) {
-            sendAndVerifyMessage(c, 100);
-        }
-        assertEquals(1, count.get());
-    }
-
-    private void sendAndVerifyMessage(Connection connection, int size) throws Exception {
-        CountDownLatch latch = new CountDownLatch(1);
-        boolean messageReceived = basicGetBasicConsume(connection, QUEUE, latch, size);
-        assertTrue(messageReceived, "Message has not been received");
-    }
-
-    private Connection basicGetBasicConsume(ConnectionFactory connectionFactory, String queue, final CountDownLatch latch)
-            throws IOException, TimeoutException {
-        Connection connection = connectionFactory.newConnection();
-        Channel channel = connection.createChannel();
-        channel.queueDeclare(queue, false, false, false, null);
-        channel.queuePurge(queue);
-
-        channel.basicPublish("", queue, null, new byte[20000]);
-
-        channel.basicConsume(queue, false, new DefaultConsumer(channel) {
-            @Override
-            public void handleDelivery(String consumerTag, Envelope envelope, AMQP.BasicProperties properties, byte[] body) throws IOException {
-                getChannel().basicAck(envelope.getDeliveryTag(), false);
-                latch.countDown();
-            }
-        });
-
-        return connection;
-    }
-
-    private boolean basicGetBasicConsume(Connection connection, String queue, final CountDownLatch latch, int msgSize)
-            throws Exception {
-        Channel channel = connection.createChannel();
-        channel.queueDeclare(queue, false, false, false, null);
-        channel.queuePurge(queue);
-
-        channel.basicPublish("", queue, null, new byte[msgSize]);
-
-        final String tag = channel.basicConsume(queue, false, new DefaultConsumer(channel) {
-
-            @Override
-            public void handleDelivery(String consumerTag, Envelope envelope, AMQP.BasicProperties properties, byte[] body) throws IOException {
-                getChannel().basicAck(envelope.getDeliveryTag(), false);
-                latch.countDown();
-            }
-        });
-
-        boolean done = latch.await(20, TimeUnit.SECONDS);
-        channel.basicCancel(tag);
-        return done;
-    }
-
-    private void safeClose(Connection connection) {
-        if (connection != null) {
-            try {
-                connection.abort();
-            } catch (Exception e) {
-                // OK
-            }
-        }
-    }
-}
