+++ src/test/resources/property-file-initialisation/configuration.properties
@@ -17,9 +17,3 @@ rabbitmq.topology.recovery.enabled=false
 rabbitmq.connection.recovery.interval=10000
 rabbitmq.channel.rpc.timeout=10000
 rabbitmq.channel.should.check.rpc.response.type=true
-rabbitmq.use.nio=true
-rabbitmq.nio.read.byte.buffer.size=32000
-rabbitmq.nio.write.byte.buffer.size=32000
-rabbitmq.nio.nb.io.threads=2
-rabbitmq.nio.write.enqueuing.timeout.in.ms=5000
-rabbitmq.nio.write.queue.capacity=1000
